Both views fully working and synchronized (#29)
* Both views fully working and synchronized

Add map legend and travel direction

* Update Map: fix clearDeliveries and simplify getNodeDeliveryManIndex

* Fix code warnings

+++ src/main/java/fr/insa/lyon/pld/agile/tsp/TSPSolverImplementation1.java
@@ -41,7 +41,7 @@ public class TSPSolverImplementation1 extends TSPSolverTemplate {
             // Update the sets of explored and unexplored nodes
             currentNode = nextNode;
             exploredNodes.add(nextNode);
-            candidateNodes.remove((Integer)nextNode);
+            candidateNodes.remove(nextNode);
         }
         
         lowerBound += edgesCosts[currentNode][0];
